Godoc reorganization and commenting.
